refactor: add task form component
